adding updates to the menus based on UX review
