+++ src/ProofOfConcepts/ActionMenus/kafkaTopicKebabMenu.tsx
@@ -7,7 +7,7 @@ import {
   DropdownGroup,
 } from "@patternfly/react-core";
 
-export class ActionsTable extends React.Component {
+export class TopicKebab extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -33,19 +33,12 @@ export class ActionsTable extends React.Component {
   render() {
     const { isOpen } = this.state;
     const dropdownItems = [
-      <DropdownGroup label="View instance information">
-        <DropdownItem>Details</DropdownItem>
-        <DropdownItem>Connection</DropdownItem>
-        <DropdownItem>Settings</DropdownItem>
-        <DropdownItem>Topics</DropdownItem>
+      <DropdownGroup>
+        <DropdownItem>Edit topic configuration</DropdownItem>
         <DropdownSeparator />
       </DropdownGroup>,
       <DropdownGroup>
-        <DropdownItem>Change instance owner</DropdownItem>
-        <DropdownSeparator />
-      </DropdownGroup>,
-      <DropdownGroup>
-        <DropdownItem>Delete instance</DropdownItem>
+        <DropdownItem>Delete topic</DropdownItem>
       </DropdownGroup>,
     ];
     return (
