--- src/ProofOfConcepts/ActionMenus/secondTopicActions.tsx
@@ -1,58 +1,0 @@
-import React from "react";
-import {
-  Dropdown,
-  DropdownToggle,
-  DropdownGroup,
-  DropdownItem,
-  DropdownSeparator,
-} from "@patternfly/react-core";
-
-export class TopicActions2 extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isOpen: false,
-    };
-    this.onToggle = (isOpen) => {
-      this.setState({
-        isOpen,
-      });
-    };
-    this.onSelect = (event) => {
-      this.setState({
-        isOpen: !this.state.isOpen,
-      });
-      this.onFocus();
-    };
-    this.onFocus = () => {
-      const element = document.getElementById("toggle-id-3");
-      element.focus();
-    };
-  }
-
-  render() {
-    const { isOpen } = this.state;
-    const dropdownItems = [
-      <DropdownGroup>
-        <DropdownItem>Edit topic configuration</DropdownItem>
-        <DropdownSeparator />
-      </DropdownGroup>,
-      <DropdownGroup>
-        <DropdownItem>Delete topic</DropdownItem>
-      </DropdownGroup>,
-    ];
-    return (
-      <Dropdown
-        onSelect={this.onSelect}
-        toggle={
-          <DropdownToggle onToggle={this.onToggle} id="toggle-id-3">
-            Actions
-          </DropdownToggle>
-        }
-        isOpen={isOpen}
-        dropdownItems={dropdownItems}
-        isGrouped
-      />
-    );
-  }
-}
